fix: New page form freezes the screen after upgrading informed (#75)
Technical notes

when the user types a character in the new page form, useEffect of useField hook is triggered, but it is not supposed to be invoked since we did not change the props. The reason why it is triggered is that we have a function prop and the references to this function are not equal during useMemo.
In order to resolve the problem, instead of passing the functions as props we pass memoized callbacks produced by useCallback.